Added a cast to the Linux implementation of Core::(POSIX)Thread - Added Maths::maximum(Uint64, Uint64)

+++ platform/source/linux/LinuxThread.cpp
@@ -30,5 +30,5 @@ using namespace Core;
 
 Uint32 Thread::Implementation::currentID()
 {
-	return syscall(SYS_gettid);
+	return static_cast<Uint32>(syscall(SYS_gettid));
 }
